+++ SharePointFramework/ProgramWebParts/package.json
@@ -1,6 +1,6 @@
 {
     "name": "pp365-program",
-    "version": "1.3.1",
+    "version": "1.4.0",
     "engines": {
         "node": ">=12.13.0"
     },
