+++ SharePointFramework/ProgramWebParts/src/webparts/programAdministration/manifest.json
