+++ SharePointFramework/ProgramWebParts/src/webparts/programBenefits/manifest.json
