+++ SharePointFramework/ProgramWebParts/src/webparts/programDeliveries/manifest.json
