+++ SharePointFramework/ProgramWebParts/src/webparts/programProjectOverview/manifest.json
