+++ SharePointFramework/ProgramWebParts/src/webparts/programRiskOverview/index.ts
@@ -1,51 +1,44 @@
 import * as ReactDom from 'react-dom'
 import { Version } from '@microsoft/sp-core-library'
 import * as strings from 'ProgramWebPartsStrings'
-import { ProgramDeliveries } from 'components/ProgramDeliveries/ProgramDeliveries'
-import { BaseProgramWebPart } from '../baseProgramWebPart/baseProgramWebPart'
+import { ProgramRiskOverview } from 'components/ProgramRiskOverview/index'
+import { BaseProgramWebPart } from '../baseProgramWebPart'
 import { IBaseWebPartComponentProps } from 'pp365-projectwebparts/lib/components/BaseWebPartComponent/types'
 import {
   IPropertyPaneConfiguration,
   PropertyPaneTextField,
   PropertyPaneToggle
 } from '@microsoft/sp-property-pane'
-import { IProgramDeliveriesProps } from 'components/ProgramDeliveries/types'
+import { IProgramRiskOverview } from 'components/ProgramRiskOverview/types'
+import { AggreationColumn } from 'models'
 
-interface IProgramDeliveriesWebPartProps extends IBaseWebPartComponentProps {
+interface IProgramRiskProps extends IBaseWebPartComponentProps {
   webPartTitle: string
   dataSource: string
   showExcelExportButton: boolean
   showSearchBox: boolean
+  columns: AggreationColumn[]
   showCommandBar: boolean
-  columns: Array<{
-    key: string
-    fieldName: string
-    name: string
-    minWidth: number
-    maxWidth: number
-    isMultiline: boolean
-    isResizable: boolean
-  }>
 }
 
-export default class ProgramDeliveriesWebPart extends BaseProgramWebPart<IProgramDeliveriesWebPartProps> {
+export default class ProgramRiskOverviewWebPart extends BaseProgramWebPart<IProgramRiskProps> {
   public async onInit(): Promise<void> {
     await super.onInit()
   }
 
   public render(): void {
-    this.renderComponent<IProgramDeliveriesProps>(ProgramDeliveries, {
-      webPartTitle: this.properties.webPartTitle,
+    this.renderComponent<IProgramRiskOverview>(ProgramRiskOverview, {
       context: this.context,
       dataAdapter: this.dataAdapter,
       properties: {
         dataSource: this.properties.dataSource,
-        showCommandBar: this.properties.showCommandBar,
-        showSearchBox: this.properties.showSearchBox,
         showExcelExportButton: this.properties.showExcelExportButton,
+        showSearchBox: this.properties.showSearchBox,
+        showCommandBar: this.properties.showCommandBar,
         columns: this.properties.columns,
         displayMode: this.displayMode
       },
+      webPartTitle: this.properties.webPartTitle,
       onUpdateProperty: this._onUpdateProperty.bind(this)
     })
   }
@@ -54,13 +47,13 @@ export default class ProgramDeliveriesWebPart extends BaseProgramWebPart<IProgra
     ReactDom.unmountComponentAtNode(this.domElement)
   }
 
-  protected get dataVersion(): Version {
-    return Version.parse('1.0')
-  }
-
-  private _onUpdateProperty(key: string, value: any) {
+  public _onUpdateProperty(key: string, value: any) {
     this.properties[key] = value
     this.context.propertyPane.refresh()
+  }
+
+  protected get dataVersion(): Version {
+    return Version.parse('1.0')
   }
 
   public getPropertyPaneConfiguration(): IPropertyPaneConfiguration {
@@ -72,11 +65,11 @@ export default class ProgramDeliveriesWebPart extends BaseProgramWebPart<IProgra
               groupFields: [
                 PropertyPaneTextField('webPartTitle', {
                   label: strings.WebPartTitleLabel,
-                  value: strings.DeliveriesTitle
+                  value: strings.RiskWebPartTitle
                 }),
                 PropertyPaneTextField('dataSource', {
                   label: strings.DataSourceLabel,
-                  value: strings.DeliveriesDatasource
+                  value: strings.RiskDataSource
                 }),
                 PropertyPaneToggle('showCommandBar', {
                   label: strings.ShowCommandBarLabel,
