+++ SharePointFramework/ProgramWebParts/src/webparts/programTimeline/manifest.json
@@ -7,7 +7,6 @@
     "manifestVersion": 2,
     "requiresCustomScript": false,
     "supportedHosts": ["SharePointWebPart"],
-
     "preconfiguredEntries": [{
         "groupId": "5c03119e-3074-46fd-976b-c60198311f70",
         "group": { "default": "Other" },
